+++ scripts/entrypoint.sh
@@ -5,14 +5,6 @@ set -e
 # FUNCTIONS
 # =========
 
-pull_shared_shib_files() {
-    # sync with existing files in target directory (mapped volume)
-    mkdir -p "$GLUU_SHIB_TARGET_DIR" "$GLUU_SHIB_SOURCE_DIR"
-    if [ -n "$(ls -A $GLUU_SHIB_TARGET_DIR 2>/dev/null)" ]; then
-        cp -r $GLUU_SHIB_TARGET_DIR/* $GLUU_SHIB_SOURCE_DIR
-    fi
-}
-
 get_debug_opt() {
     debug_opt=""
     if [ -n "${GLUU_DEBUG_PORT}" ]; then
@@ -28,35 +20,47 @@ run_wait() {
     python /app/scripts/wait.py
 }
 
-run_entrypoint() {
+move_builtin_jars() {
     # move oxtrust-api lib
     if [ ! -f /opt/gluu/jetty/identity/custom/libs/oxtrust-api-server.jar ]; then
         mkdir -p /opt/gluu/jetty/identity/custom/libs
         mv /tmp/oxtrust-api-server.jar /opt/gluu/jetty/identity/custom/libs/oxtrust-api-server.jar
     fi
+}
 
+run_entrypoint() {
     if [ ! -f /deploy/touched ]; then
         python /app/scripts/entrypoint.py
-        pull_shared_shib_files
         ln -s /etc/certs/gluu_https.crt /etc/certs/httpd.crt
         touch /deploy/touched
     fi
+}
+
+run_jca_sync() {
+    python3 /app/scripts/jca_sync.py &
+}
+
+run_mod_context() {
+    python3 /app/scripts/mod_context.py
 }
 
 # ==========
 # ENTRYPOINT
 # ==========
 
+move_builtin_jars
+
 if [ -f /etc/redhat-release ]; then
     source scl_source enable python27 && run_wait
+    source scl_source enable python3 && run_jca_sync
     source scl_source enable python27 && run_entrypoint
+    source scl_source enable python27 && run_mod_context
 else
     run_wait
+    run_jca_sync
     run_entrypoint
+    run_mod_context
 fi
-
-# monitor filesystem changes on Shibboleth-related files
-sh /app/scripts/shibwatcher.sh &
 
 # enable passport menu (a workaround for https://git.io/fjQCu)
 mkdir -p /opt/gluu/node/passport/server
@@ -67,8 +71,12 @@ mkdir -p /opt/gluu/radius && echo 'dummy file to enable Radius menu' > /opt/gluu
 # # enable shib3 menu (a workaround for https://git.io/fjQCW)
 # mkdir -p /opt/gluu/jetty/idp/webapps && echo 'dummy file to enable Shibboleth3 menu' > /opt/gluu/jetty/idp/webapps/idp.war
 
-cd /opt/gluu/jetty/identity
+# python /app/scripts/shibwatcher.py watch-files &
+# python /app/scripts/shibwatcher.py watch-containers &
+
+
 mkdir -p /opt/jetty/temp
+cd /opt/gluu/jetty/identity
 exec java \
     -server \
     -XX:+DisableExplicitGC \